+++ test/integration/group_spec.rb
@@ -33,30 +33,29 @@ describe 'Group' do
       expect(result.message).to eq('Permission read:true,modify:true on path /etc/replication set for group somegroup')
     end
 
-    # TODO: currently not working due to error 500
-    # it 'should succeed adding another group as a member' do
-    #
-    #   # ensure member group doesn't exist prior to testing
-    #   member_group = @aem.group('/home/groups/s/', 'somemembergroup')
-    #   if member_group.exists().data == true
-    #     member_group.delete()
-    #   end
-    #   result = member_group.exists()
-    #   expect(result.data).to eq(false)
-    #
-    #   # create member group
-    #   result = member_group.create()
-    #   expect(result.message).to match(/^Group somemembergroup created at \/home\/groups\/s\/.+/)
-    #
-    #   # ensure member group exists
-    #   result = member_group.exists()
-    #   expect(result.message).to match(/^Group somemembergroup exists at \/home\/groups\/s\/.+/)
-    #   expect(result.data).to eq(true)
-    #
-    #   # add user as member to the group
-    #   result = @group.add_member('somemembergroup')
-    #   expect(result.message).to eq('User/group somemembergroup added to group somegroup')
-    # end
+    it 'should succeed adding another group as a member' do
+
+      # ensure member group doesn't exist prior to testing
+      member_group = @aem.group('/home/groups/s/', 'somemembergroup')
+      if member_group.exists().data == true
+        member_group.delete()
+      end
+      result = member_group.exists()
+      expect(result.data).to eq(false)
+
+      # create member group
+      result = member_group.create()
+      expect(result.message).to match(/^Group somemembergroup created at \/home\/groups\/s\/.+/)
+
+      # ensure member group exists
+      result = member_group.exists()
+      expect(result.message).to match(/^Group somemembergroup exists at \/home\/groups\/s\/.+/)
+      expect(result.data).to eq(true)
+
+      # add user as member to the group
+      result = @group.add_member('somemembergroup')
+      expect(result.message).to eq('User/group somemembergroup added to group somegroup')
+    end
 
   end